backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
pydantic-settings>=2.0
supabase>=2.0
anthropic>=0.40.0
openai>=1.50.0
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29.0
alembic>=1.13.0
redis>=5.0
python-multipart>=0.0.9
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
Pillow>=10.0
pytest>=8.0
pytest-asyncio>=0.24.0
